requirements.txt
aiofiles==25.1.0
aiogram==3.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.22.0
alembic==1.17.2
annotated-types==0.7.0
asgiref==3.11.0
attrs==25.4.0
boundless-aiogram==1.0.7
certifi==2025.11.12
Django==6.0
dotenv==0.9.9
frozenlist==1.8.0
greenlet==3.3.0
idna==3.11
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
SQLAlchemy==2.0.45
sqlparse==0.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
yarl==1.22.0
